@@ -1,44 +1,39 @@
 {
   "name": "soap",
-  "version": "0.45.0",
+  "version": "1.4.1",
   "description": "A minimal node SOAP client",
   "engines": {
-    "node": ">=12.0.0"
+    "node": ">=14.17.0"
   },
   "author": "Vinay Pulim <v@pulim.com>",
   "dependencies": {
-    "axios-ntlm": "^1.2.0",
-    "debug": "^4.3.2",
-    "formidable": "^2.0.1",
+    "axios": "^1.12.2",
+    "axios-ntlm": "^1.4.6",
+    "debug": "^4.4.3",
+    "formidable": "^3.5.4",
     "get-stream": "^6.0.1",
     "lodash": "^4.17.21",
-    "sax": ">=0.6",
+    "sax": "^1.4.1",
     "strip-bom": "^3.0.0",
-    "uuid": "^8.3.2",
-    "whatwg-mimetype": "3.0.0",
-    "xml-crypto": "^2.1.3"
-  },
-  "peerDependencies": {
-    "axios": ">=0.21.1"
+    "whatwg-mimetype": "4.0.0",
+    "xml-crypto": "^6.1.2"
   },
   "repository": {
     "type": "git",
     "url": "https://github.com/vpulim/node-soap.git"
   },
-  "main": "./index.js",
+  "main": "./lib/soap.js",
   "types": "./lib/soap.d.ts",
   "directories": {
     "lib": "./lib",
     "test": "./test"
   },
   "scripts": {
-    "prepare": "npm run build",
     "build": "tsc -p .",
-    "clean": "rm -rf lib",
     "watch": "tsc -w -p .",
-    "lint": "tslint -p tsconfig.json",
+    "lint": "eslint .",
     "toc": "./node_modules/.bin/doctoc Readme.md --github --maxlevel 3",
-    "cover": "nyc --extension=.ts --reporter=lcov --reporter=html --reporter=text mocha --timeout 15000 --exit test/*-test.js test/security/*.js",
+    "cover": "OPENSSL_ENABLE_SHA1_SIGNATURES=1 nyc --extension=.ts --reporter=lcov --reporter=html --reporter=text mocha --no-parallel --timeout 30000 --exit test/*-test.js test/security/*.js",
     "coveralls": "cat ./coverage/lcov.info | ./node_modules/coveralls/bin/coveralls.js -v",
     "docs": "typedoc --out docs",
     "test": "mocha --timeout 15000 --bail --exit test/*-test.js test/security/*.js"
@@ -48,37 +43,33 @@
   ],
   "license": "MIT",
   "devDependencies": {
-    "@types/axios": "^0.14.0",
-    "@types/debug": "^4.1.7",
-    "@types/express": "^4.17.13",
-    "@types/formidable": "^2.0.4",
-    "@types/lodash": "^4.14.172",
-    "@types/node": "^11.15.54",
-    "@types/request": "^2.48.7",
-    "@types/sax": "^1.2.3",
-    "@types/uuid": "^8.3.1",
-    "@types/whatwg-mimetype": "^2.1.1",
-    "body-parser": "^1.15.2",
-    "colors": "^1.3.3",
-    "coveralls": "^3.1.1",
-    "diff": "^4.0.1",
-    "doctoc": "^2.0.1",
-    "duplexer": "~0.1.1",
-    "express": "^4.16.4",
-    "finalhandler": "^1.1.1",
-    "glob": "^7.1.7",
-    "jshint": "^2.13.1",
-    "mocha": "^6.1.4",
-    "nyc": "^14.1.1",
-    "readable-stream": "~2.0.2",
-    "semver": "^5.6.0",
-    "serve-static": "^1.14.1",
+    "@eslint/js": "^9.33.0",
+    "@types/debug": "^4.1.12",
+    "@types/formidable": "^3.4.5",
+    "@types/lodash": "^4.17.20",
+    "@types/node": "^14.0.0",
+    "@types/sax": "^1.2.7",
+    "@types/whatwg-mimetype": "^3.0.2",
+    "body-parser": "^2.2.0",
+    "colors": "^1.4.0",
+    "diff": "^8.0.2",
+    "doctoc": "^2.2.1",
+    "duplexer": "~0.1.2",
+    "eslint": "^9.33.0",
+    "express": "^5.1.0",
+    "finalhandler": "^2.1.0",
+    "glob": "^11.0.1",
+    "mocha": "^11.7.1",
+    "nyc": "^17.1.0",
+    "readable-stream": "^4.7.0",
+    "semver": "^7.7.1",
+    "serve-static": "^2.2.0",
     "should": "^13.2.3",
-    "sinon": "^1.17.7",
-    "source-map-support": "^0.5.10",
-    "timekeeper": "^2.1.2",
-    "tslint": "^5.18.0",
-    "typedoc": "^0.20.37",
-    "typescript": "^3.9.10"
+    "sinon": "^21.0.0",
+    "source-map-support": "^0.5.21",
+    "timekeeper": "^2.3.1",
+    "typedoc": "^0.28.10",
+    "typescript": "^5.9.2",
+    "typescript-eslint": "^8.40.0"
   }
 }
