@@ -1,18 +1,16 @@
 'use strict';
 
 var soap = require('..'),
-    http = require('http'),
-    assert = require('assert'),
-  req = require('request'),
+  assert = require('assert'),
   httpClient = require('../lib/http.js').HttpClient,
   util = require('util'),
   events = require('events'),
   createSocketStream = require('./_socketStream');
 
-it('should allow customization of httpClient, the wsdl file, and associated data download should pass through it', function(done) {
+it('should allow customization of httpClient, the wsdl file, and associated data download should pass through it', function (done) {
 
   //Make a custom http agent to use streams instead of a real net socket
-  function CustomAgent(options, wsdl, xsd){
+  function CustomAgent(options, wsdl, xsd) {
     var self = this;
     events.EventEmitter.call(this);
     self.requests = [];
@@ -25,7 +23,7 @@ it('should allow customization of httpClient, the wsdl file, and associated data
 
   util.inherits(CustomAgent, events.EventEmitter);
 
-  CustomAgent.prototype.addRequest = function(req, options) {
+  CustomAgent.prototype.addRequest = function (req, options) {
     if (/\?xsd$/.test(req.path)) {
       req.onSocket(this.xsdStream);
     } else {
@@ -34,20 +32,20 @@ it('should allow customization of httpClient, the wsdl file, and associated data
   };
 
   //Custom httpClient
-  function MyHttpClient (options, wsdlSocket, xsdSocket){
-    httpClient.call(this,options);
-    this.agent = new CustomAgent(options, wsdlSocket, xsdSocket);
+  class MyHttpClient extends httpClient {
+    constructor(options, wsdlSocket, xsdSocket) {
+      super(options);
+      this.agent = new CustomAgent(options, wsdlSocket, xsdSocket);
+    }
   }
 
-  util.inherits(MyHttpClient, httpClient);
-
-  MyHttpClient.prototype.request = function(rurl, data, callback, exheaders, exoptions) {
+  MyHttpClient.prototype.request = function (rurl, data, callback, exheaders, exoptions) {
     var self = this;
     var options = self.buildRequest(rurl, data, exheaders, exoptions);
     //Specify agent to use
     options.httpAgent = this.agent;
     var headers = options.headers;
-    var req = self._request(options).then(function(res) {
+    var req = self._request(options).then(function (res) {
       res.data = self.handleResponse(req, res, res.data);
       callback(null, res, res.data);
       if (headers.Connection !== 'keep-alive') {
@@ -60,13 +58,13 @@ it('should allow customization of httpClient, the wsdl file, and associated data
   };
 
   var httpCustomClient = new MyHttpClient({},
-    createSocketStream(__dirname + '/wsdl/xsdinclude/xsd_include_http.wsdl', 2708),
-    createSocketStream(__dirname + '/wsdl/xsdinclude/types.xsd', 982)
+    createSocketStream(__dirname + '/wsdl/xsdinclude/xsd_include_http.wsdl'),
+    createSocketStream(__dirname + '/wsdl/xsdinclude/types.xsd')
   );
   var url = 'http://localhost:50000/Dummy.asmx?wsdl';
   soap.createClient(url,
-    {httpClient: httpCustomClient},
-    function(err, client) {
+    { httpClient: httpCustomClient },
+    function (err, client) {
       assert.ok(client);
       assert.ifError(err);
       assert.equal(client.httpClient, httpCustomClient);
